mywhoosh_to_garmin/requirements.txt
garminconnect==0.2.36
playwright==1.49.1
python-dotenv==1.2.1
fitparse==1.2.0
requests==2.31.0
